refactor spans
